Add roxctl to BusyBox consolidation
Create roxctl/app package and integrate into central binary:
- Created roxctl/app/app.go with Run() function
- Updated roxctl/main.go to call app.Run()
- Moved getCommandPath() to roxctl/main_test.go for test usage
- Added roxctl case to central/main.go dispatcher
- Updated import validator to allow central to import roxctl/app
- Changed /stackrox/roxctl symlink to point to central instead of assets

Benefits:
- /stackrox/roxctl now uses consolidated binary
- Provides option to remove roxctl binaries from /assets later for space savings
- roxctl binaries in /assets still available for user downloads (for now)
