+++ image/rhel/Dockerfile
@@ -90,7 +90,7 @@ RUN mkdir -p /stackrox/bin && \
     ln -s /stackrox/central /stackrox/bin/admission-control && \
     ln -s /stackrox/central /stackrox/bin/config-controller && \
     ln -s /stackrox/central /stackrox/bin/roxagent && \
-    ln -s /assets/downloads/cli/roxctl-linux-${TARGET_ARCH} /stackrox/roxctl && \
+    ln -s /stackrox/central /stackrox/roxctl && \
     ln -s /assets/downloads/cli/roxctl-linux-amd64 /assets/downloads/cli/roxctl-linux
 
 EXPOSE 8443
