+++ image/rhel/konflux.Dockerfile
@@ -86,7 +86,7 @@ RUN GOARCH=$(uname -m) ; \
     ln -s /stackrox/central /stackrox/bin/admission-control && \
     ln -s /stackrox/central /stackrox/bin/config-controller && \
     ln -s /stackrox/central /stackrox/bin/roxagent && \
-    ln -s /assets/downloads/cli/roxctl-linux-$GOARCH /stackrox/roxctl && \
+    ln -s /stackrox/central /stackrox/roxctl && \
     ln -s /assets/downloads/cli/roxctl-linux-$GOARCH /assets/downloads/cli/roxctl-linux
 
 ARG BUILD_TAG
